requirements.txt
django-crispy-forms:1.11.1
pylint-django:2.4.2
django:3.1.7
pillow:8.0.1
